swagger

+++ src/main/java/pl/brych/hw_modul3/model/Car.java
@@ -1,17 +1,16 @@
 package pl.brych.hw_modul3.model;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 public class Car {
 
-    private long id;
+    private Integer id;
     private String mark;
     private String model;
     private String color;
 
-    public Car(long id, String mark, String model, String color) {
+    public Car(Integer id, String mark, String model, String color) {
         this.id = id;
         this.mark = mark;
         this.model = model;
